# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CyberShield India API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Profile
...
  /courses:
    get:
      operationId: getCourses
      tags: [courses]
      summary: Get all courses
      parameters:
        - name: level
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: is_free
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Courses list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoursesResponse"
  /courses/{id}/enroll:
    post:
      operationId: enrollCourse
      tags: [courses]
      summary: Enroll in a course
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Enrolled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseEnrollment"
  /courses/my-courses:
    get:
      operationId: getMyCourses
      tags: [courses]
      summary: Get user's enrolled courses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyCoursesResponse"

  # CTF
...
  /internships:
    get:
      operationId: getInternships
      tags: [internships]
      summary: Get available internships
      parameters:
        - name: org_type
          in: query
          schema:
            type: string
        - name: location_type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Internships list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternshipsResponse"
  /internships/{id}/apply:
    post:
      operationId: applyInternship
      tags: [internships]
      summary: Apply for an internship
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InternshipApplyInput"
      responses:
        "200":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternshipApplication"
  /internships/my-applications:
    get:
      operationId: getMyApplications
      tags: [internships]
      summary: Get user's internship applications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyApplicationsResponse"

  # Badges
...
  /companies/post-internship:
    post:
      operationId: postInternship
      tags: [companies]
      summary: Post a new internship
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInternshipInput"
      responses:
        "201":
          description: Internship posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Internship"
  /companies/stats:
    get:
      operationId: getCompanyStats
      tags: [companies]
      summary: Get company dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Company stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyStats"

  # RakshBot
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get user notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationsResponse"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # Dashboard stats
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    RegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, police, company]
        college_name:
          type: string
        college_city:
          type: string
        state:
          type: string
        station_name:
          type: string
        badge_number:
          type: string
        company_name:
          type: string
      required: [name, email, password, role]
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        skill_level:
          type: string
        skill_score:
          type: integer
        total_points:
          type: integer
        national_rank:
          type: integer
          nullable: true
        state_rank:
          type: integer
          nullable: true
        wallet_balance:
          type: number
        total_earned:
          type: number
        is_verified:
          type: integer
        college_name:
          type: string
          nullable: true
        college_city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        github_url:
          type: string
          nullable: true
        linkedin_url:
          type: string
          nullable: true
        station_name:
          type: string
          nullable: true
        badge_number:
          type: string
          nullable: true
        company_name:
          type: string
          nullable: true
        created_at:
          type: string
      required: [id, name, email, role]
...
    Course:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        instructor_name:
          type: string
          nullable: true
        instructor_org:
          type: string
          nullable: true
        level:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        duration_hours:
          type: number
          nullable: true
        total_lessons:
          type: integer
        is_free:
          type: integer
        price:
          type: number
        is_certified:
          type: integer
        enrolled_count:
          type: integer
        rating:
          type: number
        created_at:
          type: string
      required: [id, title, is_free, price, is_certified, enrolled_count, rating, created_at]
    CourseEnrollment:
      type: object
      properties:
        id:
          type: string
        course_id:
          type: string
        progress_percent:
          type: integer
        started_at:
          type: string
        completed_at:
          type: string
          nullable: true
      required: [id, course_id, progress_percent, started_at]
    CoursesResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        total:
          type: integer
      required: [courses, total]
    MyCourse:
      type: object
      properties:
        course:
          $ref: "#/components/schemas/Course"
        enrollment:
          $ref: "#/components/schemas/CourseEnrollment"
      required: [course, enrollment]
    MyCoursesResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: "#/components/schemas/MyCourse"
      required: [courses]
...
    Internship:
      type: object
      properties:
        id:
          type: string
        org_name:
          type: string
        org_type:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        requirements:
          type: string
          nullable: true
        duration_months:
          type: integer
          nullable: true
        location_type:
          type: string
          nullable: true
        location_city:
          type: string
          nullable: true
        location_state:
          type: string
          nullable: true
        stipend_amount:
          type: number
          nullable: true
        required_min_rank:
          type: integer
          nullable: true
        total_seats:
          type: integer
        filled_seats:
          type: integer
        status:
          type: string
        is_govt_certified:
          type: integer
        created_at:
          type: string
      required: [id, org_name, title, total_seats, filled_seats, status, is_govt_certified, created_at]
    InternshipsResponse:
      type: object
      properties:
        internships:
          type: array
          items:
            $ref: "#/components/schemas/Internship"
        total:
          type: integer
      required: [internships, total]
    InternshipApplyInput:
      type: object
      properties:
        cover_letter:
          type: string
      required: [cover_letter]
    InternshipApplication:
      type: object
      properties:
        id:
          type: string
        internship_id:
          type: string
        status:
          type: string
        applied_at:
          type: string
      required: [id, internship_id, status, applied_at]
    MyApplicationsResponse:
      type: object
      properties:
        applications:
          type: array
          items:
            type: object
            properties:
              application:
                $ref: "#/components/schemas/InternshipApplication"
              internship:
                $ref: "#/components/schemas/Internship"
            required: [application, internship]
      required: [applications]
...
    PostInternshipInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        requirements:
          type: string
        duration_months:
          type: integer
        location_type:
          type: string
          enum: [remote, onsite, hybrid]
        location_city:
          type: string
        stipend_amount:
          type: number
        total_seats:
          type: integer
        application_deadline:
          type: string
      required: [title, description, location_type, total_seats]
    CompanyStats:
      type: object
      properties:
        active_programs:
          type: integer
        total_reports_received:
          type: integer
        total_paid:
          type: number
        open_internships:
          type: integer
      required: [active_programs, total_reports_received, total_paid, open_internships]
...
    Notification:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        message:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        is_read:
          type: integer
        created_at:
          type: string
      required: [id, title, is_read, created_at]
    NotificationsResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unread_count:
          type: integer
      required: [notifications, unread_count]