# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberShield India API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  name: string;
  email: string;
  role: string;
  skill_level?: string;
  skill_score?: number;
  total_points?: number;
  national_rank?: number | null;
  state_rank?: number | null;
  wallet_balance?: number;
  total_earned?: number;
  is_verified?: number;
  college_name?: string | null;
  college_city?: string | null;
  state?: string | null;
  avatar_url?: string | null;
  bio?: string | null;
  github_url?: string | null;
  linkedin_url?: string | null;
  station_name?: string | null;
  badge_number?: string | null;
  company_name?: string | null;
  created_at?: string;
}